Improved handling of a certain apt-notify related bug
git-svn-id: http://svn.opengroupware.org/OpenGroupware.org/trunk@1614 fda58e92-02e2-0310-b1e4-b2b32738dd10

+++ WebUI/Mailer/OGoMailEditor/LSWImapMailEditor.m
@@ -1363,13 +1363,20 @@ static Class      StrClass        = nil;
   if (![_apt isNotNull])
     return;
 
-  cyc = [[_apt valueForKey:@"type"] isNotNull] ? YES : NO;
+  if (![[_apt valueForKey:@"globalID"] isNotNull]) {
+    /* BUG: this somehow happens when a conflict occurred */
+    [self errorWithFormat:
+	    @"Got no global-id for appointment object, can't delete it."];
+    return;
+  }
+  
+  cyc = [[_apt valueForKey:@"type"] isNotEmpty] ? YES : NO;
   
   [self runCommand:@"appointment::delete",
-              @"object",          _apt,
-              @"deleteAllCyclic", [NSNumber numberWithBool:cyc],
-              @"reallyDelete",    [NSNumber numberWithBool:YES],
-              nil];
+	  @"object",          _apt,
+	  @"deleteAllCyclic", [NSNumber numberWithBool:cyc],
+	  @"reallyDelete",    [NSNumber numberWithBool:YES],
+	nil];
 }
 
 - (id)cancelAndDeleteAppointment {
